Moved Spree::PromotionRulesHelper to Spree::Admin::PromotionRulesHelpers
This helper is only used in the backend and should be namespaced with
Admin like all other.
