+++ backend/spec/helpers/admin/promotion_rules_helper_spec.rb
@@ -1,0 +1,12 @@
+require 'spec_helper'
+module Spree
+  describe Spree::Admin::PromotionRulesHelper, type: :helper do
+    it "does not include existing rules in options" do
+      promotion = Spree::Promotion.new
+      promotion.promotion_rules << Spree::Promotion::Rules::ItemTotal.new
+
+      options = helper.options_for_promotion_rule_types(promotion)
+      expect(options).not_to match(/ItemTotal/)
+    end
+  end
+end
